requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
langchain-community
langchain-text-splitters
pypdf
faiss-cpu
sentence-transformers
scikit-learn
numpy
qdrant-client            # (for later, leave installed)
python-dotenv
requests
groq
collections